+++ src/simif.c
@@ -12,7 +12,7 @@ struct device * new_device(struct sim_state *s)
     return &d->device;
 }
 
-static int devices_does_match(const uint32_t *addr, const struct device *device)
+static int devices_does_match(const int32_t *addr, const struct device *device)
 {
     if (*addr <= device->bounds[1]) {
         if (*addr >= device->bounds[0]) {
@@ -25,7 +25,7 @@ static int devices_does_match(const uint32_t *addr, const struct device *device)
     }
 }
 
-int dispatch_op(void *ud, int op, uint32_t addr, uint32_t *data)
+int dispatch_op(void *ud, int op, int32_t addr, int32_t *data)
 {
     struct sim_state *s = ud;
     struct device *device = NULL;
@@ -49,7 +49,7 @@ int dispatch_op(void *ud, int op, uint32_t addr, uint32_t *data)
         // Unavailable memory returns all ones by architectural decision
         // TODO document this in the machine model
         if (op != OP_WRITE)
-            *data = 0xffffffff;
+            *data = (int32_t)0xffffffff;
 
         if (s->conf.verbose > 2) {
             fprintf(stderr, "%-5s @ 0x%08x %s due to invalid device\n",
